requirements.txt
runpod>=1.6.0
faster-whisper>=1.0.3
torch>=2.2.0
torchaudio>=2.2.0
nemo_toolkit[asr]>=1.23.0
demucs>=4.0.1
numpy>=1.24.0
soundfile>=0.12.1
librosa>=0.10.1
requests>=2.31.0
omegaconf>=2.3.0
hydra-core>=1.3.2
